show error, or save and redirect
