+++ app/views/optOut/SingleYearOptOutWarning.scala.html
@@ -14,16 +14,14 @@
  * limitations under the License.
  *@
 
-@import config.FrontendAppConfig
-@import views.html.layouts.unifiedLayout
-@import views.html.components._
-@import views.html.helpers.injected.ErrorSummary
-@import uk.gov.hmrc.govukfrontend.views.html.components._
-@import uk.gov.hmrc.govukfrontend.views.viewmodels._
-@import play.twirl.api.Html
-@import play.api.data.Form
 @import forms.optOut.ConfirmOptOutSingleTaxYearForm
 @import models.incomeSourceDetails.TaxYear
+@import play.api.data.Form
+@import play.twirl.api.Html
+@import uk.gov.hmrc.govukfrontend.views.html.components._
+@import views.html.components._
+@import views.html.helpers.injected.ErrorSummary
+@import views.html.layouts.unifiedLayout
 
 
 @this(
